requirements.txt
pandas
numpy==1.18.0
scipy==1.4.1
tqdm
scikit-learn
seaborn
torch
transformers
category_encoders
catboost
lightgbm
tensorflow
keras
mlflow
pyyaml
